Fixed issue where if delegate was called without being set caused a crash
Specifying my branch on this Podspec to get test app to pull proper
version. DO NOT MERRGE AS IS

Revert "Specifying my branch on this Podspec to get test app to pull proper"

This reverts commit a8a112d3277b1626622ce437bcdec7c662060441.

Updated Podspec properly

removing required setting

Error was that it was trying to call the method on a class not the
instance.