Removed obsolete temp context as we are performing all context toperations using the block syntax they will be executed in the default context queue set up for this ManagedObjectContext
Added code to store and read the config from the db. The db will be
updated and is retreived each time by overriding the User#config method
to return the first config found on the store.

removed manual assignment of config. Mantle should be responsible for
making the association from user to config and persisting.

Added tests

Test was a false positive

removed deletion of configs and just allow mantle to deal with the
relationships as config is dependent per user and there is no reason to
persist only one config

Removed method definition that was no longer used

Changed so that we look up user in LDUserBuilder .

findOrCreateByKey will use the existing user found with that key or
create a user and store them in the DB if one does not exist with that
key

ISSUEL Currently the testuserPersisted in the LDClientTest is failing
because of some weird test environment issue. Please look as I can use a
econd pair of eyes on this.

Updated launchspec

Fixed error from NSManagedObject instead of mantle object being
returned.

Insert into managed object context

Changed so that updating context only happens in the build.

Also removed the find or create and just let the caller of the function
create if needed.

Added config and user entities for adding the abulity to copy from
Mantle object and update Database instead of inserting a new one each
time.

Added purge users to purge the oldest users if there are more than 5

Add User and Config entity to allow for the ability to update attributes
on entity instead of inserting new one into managed object context

added purge method to delete older users

Moved save user logic to prevent multiple users being created each time.

refactored some things around saving and purging

Using unique attribute to update config instead of creating a new one

Updating to user the Mantles unique prperty keys

Update so that we assign to user before instering into managed object
context.

Removed recursive association

Added proper associations to mantle declarations

Removed explicit removal of dictionary from dictionaryValue.

Removed recursive and redundant associations

Flailing

Changed so that we persist the config instead of the user.

Pulling persisted config when build is called.

Updating mantle config

Trying to remove dup configs which seem to be created when saving user.

Something is off with mantle trying a different approach to not wind up
with orphaned configs

Updated datamodel

Oh wow. The duplicates were being caused by a nullify on the
relationship instead of a cascade delete

Delete orphaned configs

removed erroneous key that was added by accident