Changed so that we set the properties directly on the entity. Mantle does not save associations to core data.
Added check if config exists and if it does merge the dictionary values
from the server response and save the context

Using uniquing key and generating key on config from user key.

We are forcing mantle to update the config entry based ont he user key
attached.

created migration for new field

explicitly set userKey. We will get a recursive loop if we try to set
user and config/

All operations on the DatamanagerContext should have been wrapped in:

    [self.managedObjectContext performBlockAndWait:^{ }]

To ensure that they are queued up on the Context's private
NSOperationQueue.

Added blocks around all methods which are missing and verified agaist
the Shopnify app.

Update the user on the shared LDClient with the lates config from server

Code changes per comments

Normalized on self vs [DatabaseManager sharedInstance] when calling from
within the DatabaseManager file