Fix duplicate flag requests on return to foreground while polling
If the timer would have fired while the app is in the background, on return to the foreground the resume issues a flag request. The original code checked if the poll was suspended, which only happens when entering the background. By removing the logic that triggers a flag request from resumeConfigPolling, the second flag request is avoided. If the app returns to the foreground before the poll would have fired, then no flag request will be made. However, in that condition the flag request will be made at the time the timer fires.