[ch70792] Swift EventSource (#99)
* Integrating Swift eventsource.

* No longer set `-sdk` on `xcodebuild` commands for circle. There seems
to be some sort of interaction between `-sdk` or `SDKROOT` and the
implicit dependency resolution system. Since Xcode 10.2, Xcode can
find implicit dependencies in `OTHER_LINKER_FLAGS` such as when
Cocoapods links a framework with `-F`. When `SDKROOT` is not set, this
doesn't conflict with anything, but I believe the behavior of finding
the framework as an implicit dependency is not required (I think it
just resolves to the same target Cocoapods is building). If `SDKROOT`
is set, xcodebuild seems to decide that it must first resolve and
build the framework. It gives the warning "Multiple targets match
implicit dependency for linker flags '-framework LDSwiftEventSource'",
as LDSwiftEventSource generates a product by the same name
(`LDSwiftEventSource.framework`) regardless of the platform/sdk being
built. Now apparently it completely forgets about the set `SDKROOT`
(which normally seems to be fed through and used to resolve the target
being built) and decides that it should build the iOS target (no idea
why, it's not even the default `SDKROOT` for
`LDSwiftEventSource`). When `SDKROOT` is set to `macosx`, `xcodebuild`
proceeds to simultaneously build both the iOS and macOS targets (I
believe the macOS target is from Cocoapods operating correctly), then
fails as it's still actually building the iOS target with the macOS
SDK configuration. The actual build error is that the
'LDSwiftEventSource-iOS-umbrella.h' header can't find 'UIKit/UIKit.h'
as UIKit is not available on macOS.

* Update LDSwiftEventSource to 0.3.0, handle some minor todos.

* This updates the circle config to actually do a `pod install`. Removes
Gemfile that was for CocoaPods.

* Specify sources for pod spec lint.

* I think spec lint tries to test the spec file against code on the
master branch, so only run a spec lint if commit is on master.

* Allow warnings for now.