fix: #220
This PR fixes the warnings mentioned in #220, which were added to the Swift compiler [somewhat recently](https://github.com/apple/swift/pull/30218). The cause of the warning is setting default values for a Decodable objects properties. This causes subtle bugs where a given property may not be correctly decoded.

The solution proposed here is one of a few, simply settings the variable to a `var` while making its setter private. This retains the original behavior for those consuming the code, while also allowing Decodable to work properly.