Update cache to use new hashing key format (#208)
The canonical key for a new context could theoretically conflict with
the key used in previous iterations of the SDK. For example, an org
context with a key of `key` and an old user with a key of `org:key`
would have their canonical keys collide. In most places this doesn't
actually matter, but for the purposes of caching it certainly does.

There is a new method on the context struct now which will return a key
suitable for hashing purposes.

In addition to the above, while working through the code, I realized we
no longer need to include device and OS automatically, so that was
removed.