Add tombstone support (#209)
Prior to this commit, when a flag was deleted, it was actually removed
from memory. This allows for the (admittedly unlikely) situation where
an older version of a flag update would be received after this deletion,
causing the flag to become "undeleted".

By managing a light-weight placeholder, the tombstone ensures we do not
allow dead flags to resurrect unless directed to by a later version of
the flag.