fix: Trigger change listeners on identify (#261)
When `identify` is called, the SDK loads any cached flags directly into
the store. It then opens a new connection and tries to receive any flag
updates. When that payload is received, it is compared to the data
already in the store. If there are no differences, we don't trigger the
listeners, even though the data definitely changed.

To prevent this problem, we are always triggering a change notification
when identify is call. Even if the values were identify for some reason,
if the user changed, we should trigger the notification listeners.