fix: Sort private attributes for improved stable encoding (#401)
To help increase cache hits, we want to always use a stable encoding for
the context. This was done in a previous commit when we added the
`encoder.outputFormatting = [.sortedKeys]` modification.

A context's private attributes are sorted as a set, which makes no
guarantee about encoding order. To address this, we are sorting the
private attributes prior to the JSON encoding phase. As a result, we
would expect to see an increase in cache hits for customers.