feat: Update caching strategy to allow for greater cache use
Instead of indexing by the full context's hash, we are going to revert
to indexing by canonical key. The cache will store the hash alongside
the flag values.

This stored hash will be compared with the active context hash when the
cache is read. If the hashes are different, the SDK will fetch updated
values. If they haven't changed, then the SDK is free to wait until the
cache freshness has exceeded the configured polling interval.

As a result, the SDK should have a smoother transition from

    default -> last known values -> fresh values

as the context changes while also minimizing unnecessary API requests.