ci: Pin transitive dependency
The Quick\Nimble framework is targeting CwlPreconditionTesting, allowing
for an update to a new major version.

CwlPreconditionTesting released a version that references the new
visionOS OS type, which older swift versions don't understand. For now,
we are going to pin CwlPreconditionTesting to the pre-visionOS-aware
version.

We can remove this explicit dependency once we have dropped support for
older builds (like xcode 13).